Extract subscription usage footer helper
